m_fix : almost all syntax errors
