+++ src/parse/split_rdr.c
@@ -6,7 +6,7 @@
 /*   By: ghan <ghan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/18 21:31:30 by ghan              #+#    #+#             */
-/*   Updated: 2021/10/04 16:24:07 by ghan             ###   ########.fr       */
+/*   Updated: 2021/10/04 17:22:59 by ghan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,9 +57,6 @@ int	check_valid_rdr_symbols(char *s, int i)
 	if (ft_strlen(s + i) >= 2
 		&& *(s + i) == '&' && !is_charset(*(s + i + 1), "<>"))
 		return (0);
-	else if (ft_strlen(s + i) >= 2
-		&& *(s + i) == '&' && is_charset(*(s + i + 1), "<>"))
-		return (1);
 	else if (is_charset(*(s + i), "<>") || (ft_strlen(s + i) >= 2
 			&& *(s + i) == '&' && is_charset(*(s + i + 1), "<>")))
 		return (1);
