+++ src/parse/utils_syntax.c
@@ -6,7 +6,7 @@
 /*   By: ghan <ghan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/21 22:37:22 by ghan              #+#    #+#             */
-/*   Updated: 2021/10/04 16:13:01 by ghan             ###   ########.fr       */
+/*   Updated: 2021/10/04 17:47:32 by ghan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
